Subida de ruta carts y adición a app.js

+++ src/routes/products.router.js
@@ -25,13 +25,18 @@ router.post("/", (req, res)=>{
 
 router.put("/:pid", (req, res) => {
     const idProduct = parseInt(req.params.pid);
+    const product = productManager.getProductById(idProduct);
+    if(!product) return res.send({error: "Producto no encontrado"});
     const {campo, dato} = req.body;
     productManager.updateProduct(idProduct, campo, dato);
 })
 
 router.delete("/:pid", (req, res) => {
     const idProduct = parseInt(req.params.pid);
+    const product = productManager.getProductById(idProduct);
+    if(!product) return res.send({error: "Producto no encontrado"}); 
     productManager.deleteProduct(idProduct);
+    res.send("Producto eliminado exitosamente"); 
 })
 
 export default router;
